refactor(flashocc2): 重构 flashocc2 模型代码
- 更新 flashocc2 模型配置和训练流程
- 移除未使用的代码和导入
- 调整数据处理和模型结构
- 更新安装说明和数据准备步骤

+++ flashocc2/core/evaluation/occ_metrics.py
@@ -41,7 +41,7 @@ def pcolor(string, color, on_color=None, attrs=None):
 
 
 def getCellCoordinates(points, voxelSize):
-    return (points / voxelSize).astype(np.int)
+    return (points / voxelSize).astype(np.int32)
 
 
 def getNumUniqueCells(cells):
